- No functional changes in regards to majority of the code - Start partial refactor of sdlmain.cpp output code - Move general definitions into sdlmain.h (incomplete) - Move little part of OpenGL output related code out - Move almost all of surface output related code out - Generalize xBRZ postscaling operations into their own respective function to reuse
- Bugfix: commits 3a8d77f and ebdfc33 break intentional output window fill-in for xBRZ
- Minor: non-ppl runs can be done without granularity based loop, adjusted

- TODO: continue factoring out output code
- TODO: SDL2 build is probably not xBRZ-capable now
- TODO: see what can be used instead of ppl taskgroups for other platforms, xBRZ is too slow without parallel processing
- TODO: adjust other build configurations to include new source files / headers
