+++ include/render.h
@@ -31,12 +31,6 @@
 //Enable this for scalers to support 0 input for empty lines
 //#define RENDER_NULL_INPUT
 
-enum SCREEN_TYPES {
-	SCREEN_SURFACE,
-	SCREEN_OPENGL,
-	SCREEN_DIRECT3D
-};
-
 enum ASPECT_MODES {
     ASPECT_FALSE = 0
     ,ASPECT_TRUE
