+++ src/output/output_direct3d.h
@@ -1,0 +1,4 @@
+#ifndef DOSBOX_OUTPUT_DIRECT3D_H
+#define DOSBOX_OUTPUT_DIRECT3D_H
+
+#endif
